fix: shorten absolute paths in workflow run agent activity view (#440)
Extract shorten_paths() to a shared TUI helpers module and apply it in
render_step_agent_activity() so workflow agent events display {worktree}
and ~ prefixes instead of raw absolute paths.
